Abstract Parser Helper Functions
